+++ services/fishsense-api-workflow-worker/src/fishsense_api_workflow_worker/worker.py
@@ -50,9 +50,6 @@ from fishsense_api_workflow_worker.activities.get_dives_with_complete_laser_labe
 from fishsense_api_workflow_worker.activities.get_headtail_label_studio_project_ids_activity import (  # pylint: disable=line-too-long
     get_headtail_label_studio_project_ids_activity,
 )
-from fishsense_api_workflow_worker.activities.get_label_studio_projects_activity import (
-    get_label_studio_projects_activity,
-)
 from fishsense_api_workflow_worker.activities.get_laser_label_studio_project_ids_activity import (
     get_laser_label_studio_project_ids_activity,
 )
@@ -134,9 +131,6 @@ from fishsense_api_workflow_worker.activities.sync_users_label_studio_activity i
 from fishsense_api_workflow_worker.activities.update_dive_image_groups_activity import (
     update_dive_image_groups_activity,
 )
-from fishsense_api_workflow_worker.activities.write_dashboard_config_activity import (
-    write_dashboard_config_activity,
-)
 from fishsense_api_workflow_worker.config import configure_logging, settings
 from fishsense_api_workflow_worker.workflows.create_dive_slate_label_studio_project_workflow import (  # pylint: disable=line-too-long
     CreateDiveSlateLabelStudioProjectWorkflow,
@@ -194,9 +188,6 @@ from fishsense_api_workflow_worker.workflows.sync_label_studio_laser_labels_work
 )
 from fishsense_api_workflow_worker.workflows.sync_label_studio_species_labels_workflow import (  # pylint: disable=line-too-long
     SyncLabelStudioSpeciesLabelsWorkflow,
-)
-from fishsense_api_workflow_worker.workflows.update_dashboard_config_workflow import (
-    UpdateDashboardConfigWorkflow,
 )
 from fishsense_api_workflow_worker.workflows.update_dive_image_groups_workflow import (
     UpdateDiveImageGroupsWorkflow,
@@ -290,14 +281,6 @@ async def schedule_workflows(client: Client):
                     client,
                     "sync-label-studio-species-labels-workflow-schedule",
                     SyncLabelStudioSpeciesLabelsWorkflow,
-                    timedelta(hours=1),
-                )
-            )
-            tg.create_task(
-                schedule_workflow(
-                    client,
-                    "update-dashboard-config-workflow-schedule",
-                    UpdateDashboardConfigWorkflow,
                     timedelta(hours=1),
                 )
             )
@@ -408,7 +391,6 @@ async def main():
                 SyncLabelStudioHeadTailLabelsWorkflow,
                 SyncLabelStudioDiveSlateLabelsWorkflow,
                 SyncLabelStudioSpeciesLabelsWorkflow,
-                UpdateDashboardConfigWorkflow,
                 CreateLaserLabelStudioProjectWorkflow,
                 CreateSpeciesLabelStudioProjectWorkflow,
                 CreateHeadTailLabelStudioProjectWorkflow,
@@ -428,7 +410,6 @@ async def main():
             ],
             activity_executor=executor,
             activities=[
-                get_label_studio_projects_activity,
                 get_laser_label_studio_project_ids_activity,
                 get_headtail_label_studio_project_ids_activity,
                 get_dive_slate_label_studio_project_ids_activity,
@@ -439,7 +420,6 @@ async def main():
                 sync_dive_slate_labels_for_label_studio_project_activity,
                 sync_species_labels_for_label_studio_project_activity,
                 sync_users_label_studio_activity,
-                write_dashboard_config_activity,
                 create_laser_label_studio_project_activity,
                 create_species_label_studio_project_activity,
                 create_headtail_label_studio_project_activity,